+++ src/plugin.rs
@@ -75,9 +75,13 @@ pub fn component_animator_system<T: Component>(
 ) {
     for (entity, ref mut target, ref mut animator) in query.iter_mut() {
         if animator.state != AnimatorState::Paused {
-            animator
-                .tweenable_mut()
-                .tick(time.delta(), target, entity, &mut event_writer);
+            let speed = animator.speed();
+            animator.tweenable_mut().tick(
+                time.delta().mul_f32(speed),
+                target,
+                entity,
+                &mut event_writer,
+            );
         }
     }
 }
@@ -97,10 +101,14 @@ pub fn asset_animator_system<T: Asset>(
 ) {
     for (entity, ref mut animator) in query.iter_mut() {
         if animator.state != AnimatorState::Paused {
+            let speed = animator.speed();
             if let Some(target) = assets.get_mut(&animator.handle()) {
-                animator
-                    .tweenable_mut()
-                    .tick(time.delta(), target, entity, &mut event_writer);
+                animator.tweenable_mut().tick(
+                    time.delta().mul_f32(speed),
+                    target,
+                    entity,
+                    &mut event_writer,
+                );
             }
         }
     }